testing/web-platform/tests/old-tests/submission/Infraware/Forms/contents/Forms/datalist_options.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p>
    <h3>datalist_options</h3>
  </p>
  
  <hr>

  <div id="log"></div>

  <form method="post"
      enctype="application/x-www-form-urlencoded"
      action=""
      name="input_form">
  <p><input type='text' id='input_url' list="urls"></p>
  <datalist id="urls">
    <option value="http://www.google.com/" label="Google">
    <option value="http://www.reddit.com/" label="Reddit">
  </datalist>
  </form>

  <script>
    
    var datalist = document.getElementById("urls");
    var options = datalist.options;

    if (typeof(options) == "object") {
      test(function() {
        assert_equals(options.length, 2, "list attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("list attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
